\name{bsts-package}
\alias{bsts-package}
\alias{bsts}
\docType{package}
\title{
\packageTitle{bsts}
}

\description{
\packageDescription{bsts}
}

\details{
  \subsection{Installation note for Linux users}{ If you are installing bsts
  using \code{\link{install.packages}} on a Linux machine (and thus
  compiling yourself) you will almost certainly want to set the
  \code{Ncpus} argument to a large number.  Windows and Mac users can
  ignore this advice.  }

The DESCRIPTION file:
\packageDESCRIPTION{bsts}

\packageIndices{bsts}
}

\author{
\packageAuthor{bsts}

Maintainer: \packageMaintainer{bsts}
}

\references{
  Please see the references in the help page for the \code{\link{bsts}} function.
}

\keyword{ package }

\seealso{
  See the examples in the \code{\link{bsts}} function.
}
