\name{named.holidays}
\docType{data}
\alias{named.holidays}
\title{Holidays Recognized by Name}
\description{A character vector listing the names of pre-specified holidays.}
\value{
"NewYearsDay"                 "SuperBowlSunday"            
"MartinLutherKingDay"         "PresidentsDay"              
"ValentinesDay"               "SaintPatricksDay"           
"USDaylightSavingsTimeBegins" "USDaylightSavingsTimeEnds"  
"EasterSunday"                "USMothersDay"               
"IndependenceDay"             "LaborDay"                   
"ColumbusDay"                 "Halloween"                  
"Thanksgiving"                "MemorialDay"                
"VeteransDay"                 "Christmas"
}
\usage{
named.holidays
}

