\name{rsxfs}
\docType{data}
\alias{rsxfs}
\alias{RSXFS}
\alias{retail.sales}
\title{Retail sales, excluding food services}
\description{A monthly time series of retail sales in the US, excluding
  food services.  In millions of dollars.  Seasonally adjusted.}
\usage{
data(rsxfs)
}
\format{zoo time series}
\source{FRED.  See http://research.stlouisfed.org/fred2/series/RSXFS}
\examples{
data(rsxfs)
plot(rsxfs)
}

\keyword{datasets}
