\name{shark}
\docType{data}
\alias{shark}
\title{Shark Attacks in Florida.}
\description{An annual time series of shark attacks and fatalities in Florida.}
\usage{
data(shark)
}
\format{zoo time series}
\source{
  From Jeffrey Simonoff "Analysis of Categorical Data".
  http://people.stern.nyu.edu/jsimonof/AnalCatData/Data/Comma_separated/floridashark.csv}
\examples{
data(shark)
head(shark)
}

\keyword{datasets}
