\name{turkish}
\docType{data}
\alias{turkish}
\title{Turkish Electricity Usage}

\description{A daily time series of electricity usaage in Turkey.  }

\usage{
data(turkish)
}
\format{zoo time series}
\source{
https://robjhyndman.com/data/turkey_elec.csv
}
\examples{
data(turkish)
plot(turkish)
}

\seealso{
  \code{\link{bsts}}
}

\keyword{datasets}
