\name{bsts-package}
\alias{bsts-package}
\docType{package}
\title{
bsts
}

\description{
  Time series regression using dynamic linear models fit using
  MCMC. See Scott and Varian (2014) <DOI:10.1504/IJMMNO.2014.059942>, among many
  other sources.
}

\details{
  \subsection{Installation note for Linux users}{ If you are installing bsts
  using \code{\link{install.packages}} on a Linux machine (and thus
  compiling yourself) you will almost certainly want to set the
  \code{Ncpus} argument to a large number.  Windows and Mac users can
  ignore this advice.  }

}

\author{
Author: Steven L. Scott <steve.the.bayesian@gmail.com>
Maintainer: Steven L. Scott <steve.the.bayesian@gmail.com>
}

\references{
  Please see the references in the help page for the \code{\link{bsts}} function.
}

\keyword{ package }

\seealso{
  See the examples in the \code{\link{bsts}} function.
}
